BFormText fixes
