+++ src/components/index.ts
@@ -43,7 +43,7 @@ import BFormFloatingLabel from './BFormFloatingLabel.vue'
 import BFormSelect from './BFormSelect/BFormSelect.vue'
 import BFormSelectOption from './BFormSelect/BFormSelectOption.vue'
 import BFormSelectOptionGroup from './BFormSelect/BFormSelectOptionGroup.vue'
-import BFormText from './BFormText.vue'
+import BFormText from './BForm/BFormText.vue'
 import BFormValidFeedback from './BForm/BFormValidFeedback.vue'
 import BIcon from './BIcon.vue'
 import BImg from './BImg.vue'
