+++ src/types/TextColorVariant.d.ts
@@ -1,0 +1,17 @@
+type TextColorVariant =
+  | 'primary'
+  | 'secondary'
+  | 'success'
+  | 'danger'
+  | 'warning'
+  | 'info'
+  | 'light'
+  | 'dark'
+  | 'white'
+  | 'body'
+  | 'muted'
+  | 'black-50'
+  | 'white-50'
+  | 'reset'
+
+export default TextColorVariant
