+++ src/types/index.ts
@@ -16,4 +16,5 @@ export {default as Size} from './Size'
 export {default as SpinnerType} from './SpinnerType'
 export {default as TableField} from './TableField'
 export {default as TableItem} from './TableItem'
+export {default as TextColorVariant} from './TextColorVariant'
 export {default as VerticalAlign} from './VerticalAlign'
